Attempt #2
When doing the Statistical Analysis I was including the Canceled projects together with the Failed ones for the Unsuccessful group. I believe that was the cause of the wrong measures for Mean, Median, Variance and Standard Deviation. This new attempt was done by taking out the Canceled projects of the Unsuccessful group.
I am over the deadline for turning in this homework, but can I still receive the feedback for this new attempt for leaning purposes?
Thank you!